+++ CHANGES.md
@@ -5,6 +5,7 @@
   * Removed `Slop::Options#to_hash` continue using `Slop#to_hash` directly.
     This method also now returns symbols by default instead of strings. If
     you want strings use `opts.to_hash(false)`
+  * Removed `Slop::Options#to_help` and merged its contents into `Slop#help`
 * Added Slop::Error subclassing from StandardError which all exception
   classes should inherit from
 * Added Slop::MissingOptionError and `:required` option to Slop::Option.